requirements.txt
# Core
torch>=2.0.0
torch-geometric>=2.4.0
lightning>=2.0.0
torchmetrics>=1.0.0

# Config & Logging
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.15.0

# Tensor ops
einops>=0.6.0
einx>=0.1.0
jaxtyping>=0.2.0

# Synthetic data
botorch>=0.9.0
numpy>=1.24.0
pandas>=2.0.0
